+++ src/containerapp/azext_containerapp/_help.py
@@ -780,12 +780,12 @@ helps['containerapp ingress access-restriction list'] = """
 
 helps['containerapp ingress sticky-sessions'] = """
     type: group
-    short-summary: Commands to set Sticky affinity for a container app .
+    short-summary: Commands to set Sticky session affinity for a container app.
 """
 
 helps['containerapp ingress sticky-sessions set'] = """
     type: command
-    short-summary: Configure Sticky Affinity for a container app.
+    short-summary: Configure Sticky session for a container app.
     examples:
     - name: Set affinity to sticky for a container app.
       text: |