+++ src/containerapp/azext_containerapp/_models.py
@@ -131,10 +131,6 @@ TrafficWeight = {
     "latestRevision": False
 }
 
-StickySessions = {
-    "affinity": None
-}
-
 BindingType = {
 
 }